moved inputattributes const from familyMemberdetails to utils
