+++ src/app/components/utils/familyMemberDetailsUtils.js
@@ -1,0 +1,25 @@
+export const getInputAttributesForFamilyMemberDetails = (field) => {
+  const { fieldKey, name } = field;
+  const memberDetailsInputAttributes = [
+    {
+      name: [name, "firstName"],
+      label: "First Name",
+      fieldKey: [fieldKey, "firstName"],
+      rules: [
+        {
+          required: true,
+          message: "Please input first name of family member!",
+        },
+      ],
+    },
+    {
+      name: [name, "lastName"],
+      label: "Last Name",
+      fieldKey: [fieldKey, "lastName"],
+      rules: [
+        { required: true, message: "Please input last name of family member!" },
+      ],
+    },
+  ];
+  return memberDetailsInputAttributes;
+};
